Create grid, buttons, and generate functionality
Built a 16 x 16 grid of squares, and each square turns black once it has
been hovered over. Buttons are in place but only then generate button is
functional. It takes the value of the number input and creates a new
grid of that size. Input validation is in place for grids too large or
too small.