Create grid, controls, and generation functionality
Built a 16 x 16 grid of squares where each square turns black after it
has been hovered over. Added controls. The only working button is for
generating a new grid, which takes the value in the input box and
creates a grid of that dimension. Input validation in place for grids
too small or too large.